kiwoom/kiwoom_mvp/requirements.txt
# 키움증권 주식 데이터 수집기 - Python 패키지 의존성
# Phase 1 MVP 요구사항

# GUI 및 이벤트 처리 (키움 API 필수)
PyQt5==5.15.7
pywin32==306

# 데이터 처리 및 분석
pandas>=1.5.0
numpy>=1.21.0

# 차트 생성 및 시각화
plotly>=5.0.0
kaleido>=0.2.1

# CSV 및 파일 처리
openpyxl>=3.0.0
chardet>=4.0.0

# 환경변수 관리
python-dotenv>=0.19.0

# 로깅 및 유틸리티
colorama>=0.4.4
tqdm>=4.62.0

# 날짜 및 시간 처리
python-dateutil>=2.8.0

# HTTP 요청 (추가 데이터 수집용)
requests>=2.25.0